+++ src/Types.php
@@ -14,6 +14,7 @@ namespace BEAR\AppMeta;
  * @psalm-type Context = non-empty-string
  * @psalm-type AppDir = non-empty-string
  * @psalm-type TmpDir = non-empty-string
+ * @psalm-type ScriptDir = non-empty-string
  * @psalm-type LogDir = non-empty-string
  * @psalm-type UriPath = non-empty-string
  * @psalm-type FilePath = non-empty-string